requirements.txt
pyqt5
numpy
tensorflow-gpu==2.2.0
keras
opencv-python